python/qmp: clear events on get_events() call
All callers in the tree *already* clear the events after a call to
get_events(), do it automatically and update callers to know that they
don't need to do it manually anymore.

These semantics are *quite* a bit easier to emulate with Async QMP, and
nobody appears to be abusing some emergent properties of what happens if
you decide not to clear them, so let's dial down to the dumber, simpler
thing.

(Specifically: callers of clear() right after a call to get_events() are
more likely expressing their desire to not see an event they've
witnessed twice, whereas callers of clear_events() not in relation to a
recent call to pull_event/get_events are likely expressing their desire
to simply drop *all* pending events straight onto the floor.)

Signed-off-by: John Snow <jsnow@redhat.com>