python, iotests: replace qmp with aqmp.sync_qmp
There's a few rough edges to make it work, but hey!

A few hacks to prove this works:

1) Disable the scary "do not use" warning, which interferes with test
output

2) Disable the aqmp logger, which likes to (at the moment) print out
intermediate warnings and errors that cause session termination; disable
them so they don't interfere with the job output. Leave any "CRITICAL"
warnings enabled though, those are ones that we should never see, no
matter what.

3) The Async QMP session does not know if EOFError is actually a problem
or not. When issuing a QUIT command, the remote will hang up and Async
QMP will get an EOF, which it dutifully reports back out upon calling
disconnect(). AQMP is "vocabulary agnostic" and I do not intend to
change that. Modify the QEMU Machine class instead to dismiss EOF as not
a problem when it calls close on the QMP object.

Signed-off-by: John Snow <jsnow@redhat.com>